@@ -20,28 +20,26 @@ class ValueFunction;
 // Base value class representing a nuua value.
 class Value
 {
+    void build_from_type(Type *type);
     public:
         // The type of the value.
         Type type;
-        bool deallocated = false;
         // Using a union to avoid unessesary memory.
         union {
             // Stores the reporesentation of the VALUE_INT.
-            int64_t value_int = 0;
+            int64_t value_int = NULL;
             // Stores the representation of the VALUE_FLOAT.
             double value_float;
             // Stores the representation of the VALUE_BOOL.
             bool value_bool;
             // Stores the representation of the VALUE_STRING.
-            std::string *value_string;
+            std::unique_ptr<std::string> value_string;
             // Stores the representation of the VALUE_LIST.
-            std::vector<Value> *value_list;
+            std::unique_ptr<std::vector<Value>> value_list;
             // Stores the representation of the VALUE_DICT.
-            ValueDictionary *value_dict;
+            std::unique_ptr<ValueDictionary> value_dict;
             // Stores the representation of the VALUE_FUN.
-            ValueFunction *value_fun;
-            // Stores the ref representation.
-            Value *value_ref;
+            std::unique_ptr<ValueFunction> value_fun;
         };
         // The following are the basic constructors for the value. Each one respresents
         // a diferent value to be stored. They pretty much speak by themselves.
@@ -57,21 +55,21 @@ class Value
             : type({ VALUE_BOOL }), value_bool(a) {}
         // String value.
         Value(std::string &a)
-            : type({ VALUE_STRING }), value_string(new std::string(a)) {}
+            : type({ VALUE_STRING }), value_string(std::make_unique<std::string>(a)) {}
         // List value.
-        Value(std::vector<Value> &a)
-            : type({ VALUE_LIST }), value_list(new std::vector<Value>(a)) {}
+        Value(std::vector<Value> &a, std::shared_ptr<Type> &inner_type)
+            : type({ VALUE_LIST, inner_type }), value_list(std::make_unique<std::vector<Value>>(a)) {}
         // The following two constructors are basically defined in the value.cpp since
         // They make use of a forward declared constructor.
-        Value(std::unordered_map<std::string, Value> &a, std::vector<std::string> &b);
+        Value(std::unordered_map<std::string, Value> &a, std::vector<std::string> &b, std::shared_ptr<Type> &inner_type);
         Value(size_t index, registers_size_t registers);
         // Create default initialized value, given the type.
-        Value(Type *type);
-        ~Value();
+        Value(std::shared_ptr<Type> &type);
+        Value(Type &type);
+        Value(const Value &value);
+        ~Value() {}
         // Copies the current value to the destnation.
-        void copy_to(Value *dest);
-        // Deallocates the space used by the value (frees the memory of the current value).
-        void deallocate();
+        void copy_to(Value *dest) const;
 };
 
 // Defines how a dictionary value is.
@@ -85,9 +83,9 @@ class ValueDictionary
         // important.
         std::vector<std::string> key_order;
         // Adds a new element to the dictionary.
-        void insert(std::string &key, Value value);
+        void insert(const std::string &key, const Value &value);
         // The basic constructor of the dictionary.
-        ValueDictionary(std::unordered_map<std::string, Value> values, std::vector<std::string> key_order)
+        ValueDictionary(const std::unordered_map<std::string, Value> &values, const std::vector<std::string> &key_order)
             : values(values), key_order(key_order) {}
 };
 
