@@ -26,11 +26,11 @@
 #define ADD_LOG_PAR(line, col, msg) logger->add_entity(this->file, line, col, msg)
 #define EXPECT_NEW_LINE() if (!this->match_any({{ TOKEN_NEW_LINE, TOKEN_EOF }})) { \
     ADD_LOG("Expected a new line or EOF but got '" + CURRENT().to_string() + "'."); exit(logger->crash()); }
-#define NEW_NODE(type, ...) (new type(this->file, PLINE(), PCOL(), __VA_ARGS__))
+#define NEW_NODE(type, ...) (std::make_shared<type>(this->file, PLINE(), PCOL(), __VA_ARGS__))
 
 // Stores the parsing file stack, to avoid
 // cyclic imports.
-static std::vector<const std::string *> file_stack;
+static std::vector<std::shared_ptr<const std::string>> file_stack;
 
 #define PREVENT_CYCLIC(file_ptr) \
 { \
@@ -44,16 +44,16 @@ static std::vector<const std::string *> file_stack;
 // parsed Abstract Syntax Tree and the pointer to the
 // original long lived file name string.
 // Acts as a temporal cache to avoid re-parsing a file.
-static std::unordered_map<std::string, std::pair<std::vector<Statement *> *, const std::string *>> parsed_files;
+static std::unordered_map<std::string, std::pair<std::shared_ptr<std::vector<std::shared_ptr<Statement>>>, std::shared_ptr<const std::string>>> parsed_files;
 
-Token *Parser::consume(TokenType type, const char *message)
+Token *Parser::consume(const TokenType type, const std::string &message)
 {
     if (this->current->type == type) return NEXT();
-    ADD_LOG(std::string(message));
+    ADD_LOG(message);
     exit(logger->crash());
 }
 
-bool Parser::match(TokenType token)
+bool Parser::match(const TokenType token)
 {
     if (CHECK(token)) {
         if (token != TOKEN_EOF) NEXT();
@@ -62,9 +62,9 @@ bool Parser::match(TokenType token)
     return false;
 }
 
-bool Parser::match_any(std::vector<TokenType>  tokens)
+bool Parser::match_any(const std::vector<TokenType> &tokens)
 {
-    for (TokenType &token : tokens) {
+    for (const TokenType &token : tokens) {
         if (CHECK(token)) {
             if (token != TOKEN_EOF) NEXT();
             return true;
@@ -86,7 +86,7 @@ primary -> "false"
     | DICTIONARY
     | "(" expression ")"
 */
-Expression *Parser::primary()
+std::shared_ptr<Expression> Parser::primary()
 {
     if (this->match(TOKEN_FALSE)) return NEW_NODE(Boolean, false);
     if (this->match(TOKEN_TRUE)) return NEW_NODE(Boolean, true);
@@ -95,21 +95,21 @@ Expression *Parser::primary()
     if (this->match(TOKEN_STRING)) return NEW_NODE(String, PREVIOUS().to_string());
     if (this->match(TOKEN_IDENTIFIER)) return NEW_NODE(Variable, PREVIOUS().to_string());
     if (this->match(TOKEN_LEFT_SQUARE)) {
-        std::vector<Expression *> values;
+        std::vector<std::shared_ptr<Expression> > values;
         if (this->match(TOKEN_RIGHT_SQUARE)) return NEW_NODE(List, values);
         for (;;) {
             if (IS_AT_END()) {
                 ADD_LOG("Unfinished list, Expecting ']' after the last list element.");
                 exit(logger->crash());
             }
-            values.push_back(expression());
+            values.push_back(std::move(expression()));
             if (this->match(TOKEN_RIGHT_SQUARE)) break;
             this->consume(TOKEN_COMMA, "Expected ',' after a list element");
         }
         return NEW_NODE(List, values);
     }
     if (this->match(TOKEN_LEFT_BRACE)) {
-        std::unordered_map<std::string, Expression *> values;
+        std::unordered_map<std::string, std::shared_ptr<Expression>> values;
         std::vector<std::string> keys;
         if (this->match(TOKEN_RIGHT_BRACE)) return NEW_NODE(Dictionary, values, keys);
         for (;;) {
@@ -117,13 +117,13 @@ Expression *Parser::primary()
                 ADD_LOG("Unfinished dictionary, Expecting '}' after the last dictionary element.");
                 exit(logger->crash());
             }
-            Expression *key = this->expression();
+            std::shared_ptr<Expression> key = this->expression();
             if (key->rule != RULE_VARIABLE) {
                 ADD_LOG("Expected an identifier as a key");
                 exit(logger->crash());
             }
             this->consume(TOKEN_COLON, "Expected ':' after dictionary key");
-            std::string name = static_cast<Variable *>(key)->name;
+            std::string name = std::static_pointer_cast<Variable>(key)->name;
             values[name] = this->expression();
             keys.push_back(name);
             if (this->match(TOKEN_RIGHT_BRACE)) break;
@@ -132,20 +132,20 @@ Expression *Parser::primary()
         return NEW_NODE(Dictionary, values, keys);
     }
     if (this->match(TOKEN_LEFT_PAREN)) {
-        Expression *value = this->expression();
+        std::shared_ptr<Expression> value = this->expression();
         this->consume(TOKEN_RIGHT_PAREN, "Expected ')' after a group expression");
         return NEW_NODE(Group, value);
     }
     /*
     if (this->match(TOKEN_STICK)) {
-        std::vector<Statement *> parameters = this->parameters();
+        std::vector<std::shared_ptr<Statement> > parameters = this->parameters();
         if (!this->match(TOKEN_STICK)) {
             parameters = this->parameters();
             this->consume(TOKEN_STICK, "Expected '|' after the closure parameters");
         }
         std::string return_type;
         if (this->match(TOKEN_COLON)) return_type = this->type(false);
-        std::vector<Statement *> body;
+        std::vector<std::shared_ptr<Statement> > body;
         if (this->match(TOKEN_RIGHT_ARROW)) {
             body.push_back(NEW_NODE(Return, this->expression()));
         } else if (this->match(TOKEN_BIG_RIGHT_ARROW)) {
@@ -170,16 +170,16 @@ unary_postfix -> primary ("[" expression "]" | slice | "(" arguments? ")")*;
 slice -> "[" expression? ":" expression? (":" expression?)? "]"
 arguments -> expression ("," expression)*;
 */
-Expression *Parser::unary_postfix()
+std::shared_ptr<Expression> Parser::unary_postfix()
 {
-    Expression *result = this->primary();
+    std::shared_ptr<Expression> result = this->primary();
     while (this->match_any({{ TOKEN_LEFT_SQUARE, TOKEN_LEFT_PAREN }})) {
         Token op = PREVIOUS();
         switch (op.type) {
             case TOKEN_LEFT_SQUARE: {
-                Expression *start = nullptr;
-                Expression *end = nullptr;
-                Expression *step = nullptr;
+                std::shared_ptr<Expression> start = nullptr;
+                std::shared_ptr<Expression> end = nullptr;
+                std::shared_ptr<Expression> step = nullptr;
                 if (this->match(TOKEN_COLON)) goto parser_is_slice1;
                 start = this->expression();
                 if (this->match(TOKEN_COLON)) {
@@ -203,7 +203,7 @@ Expression *Parser::unary_postfix()
                 break;
             }
             case TOKEN_LEFT_PAREN: {
-                std::vector<Expression *> arguments = this->arguments();
+                std::vector<std::shared_ptr<Expression> > arguments = this->arguments();
                 this->consume(TOKEN_RIGHT_PAREN, "Expected ')' after function arguments");
                 result = NEW_NODE(Call, result, arguments);
                 break;
@@ -221,11 +221,11 @@ Expression *Parser::unary_postfix()
 unary_prefix -> ("!" | "+" | "-") unary_prefix
     | unary_postfix;
 */
-Expression *Parser::unary_prefix()
+std::shared_ptr<Expression> Parser::unary_prefix()
 {
     if (this->match_any({{ TOKEN_BANG, TOKEN_PLUS, TOKEN_MINUS }})) {
         Token op = PREVIOUS();
-        Expression *expr = this->unary_prefix();
+        std::shared_ptr<Expression> expr = this->unary_prefix();
         return NEW_NODE(Unary, op, expr);
     }
     return this->unary_postfix();
@@ -234,11 +234,11 @@ Expression *Parser::unary_prefix()
 /*
 cast -> unary_prefix ("as" type)*;
 */
-Expression *Parser::cast()
+std::shared_ptr<Expression> Parser::cast()
 {
-    Expression *result = this->unary_prefix();
+    std::shared_ptr<Expression> result = this->unary_prefix();
     while (this->match(TOKEN_AS)) {
-        Type *type = this->type();
+        std::shared_ptr<Type> type = this->type();
         result = NEW_NODE(Cast, result, type);
     }
     return result;
@@ -247,12 +247,12 @@ Expression *Parser::cast()
 /*
 multiplication -> cast (("/" | "*") cast)*;
 */
-Expression *Parser::multiplication()
+std::shared_ptr<Expression> Parser::multiplication()
 {
-    Expression *result = this->cast();
+    std::shared_ptr<Expression> result = this->cast();
     while (this->match_any({{ TOKEN_SLASH, TOKEN_STAR, TOKEN_PERCENT }})) {
         Token op = PREVIOUS();
-        Expression *expr = this->cast();
+        std::shared_ptr<Expression> expr = this->cast();
         result = NEW_NODE(Binary, result, op, expr);
     }
     return result;
@@ -261,12 +261,12 @@ Expression *Parser::multiplication()
 /*
 addition -> multiplication (("-" | "+") multiplication)*;
 */
-Expression *Parser::addition()
+std::shared_ptr<Expression> Parser::addition()
 {
-    Expression *result = this->multiplication();
+    std::shared_ptr<Expression> result = this->multiplication();
     while (this->match_any({{ TOKEN_MINUS, TOKEN_PLUS }})) {
         Token op = PREVIOUS();
-        Expression *expr = this->multiplication();
+        std::shared_ptr<Expression> expr = this->multiplication();
         result = NEW_NODE(Binary, result, op, expr);
     }
     return result;
@@ -275,12 +275,12 @@ Expression *Parser::addition()
 /*
 comparison -> addition ((">" | ">=" | "<" | "<=") addition)*;
 */
-Expression *Parser::comparison()
+std::shared_ptr<Expression> Parser::comparison()
 {
-    Expression *result = this->addition();
+    std::shared_ptr<Expression> result = this->addition();
     while (this->match_any({{ TOKEN_HIGHER, TOKEN_HIGHER_EQUAL, TOKEN_LOWER, TOKEN_LOWER_EQUAL }})) {
         Token op = PREVIOUS();
-        Expression *expr = this->addition();
+        std::shared_ptr<Expression> expr = this->addition();
         result = NEW_NODE(Binary, result, op, expr);
     }
     return result;
@@ -289,12 +289,12 @@ Expression *Parser::comparison()
 /*
 equality -> comparison (("!=" | "==") comparison)*;
 */
-Expression *Parser::equality()
+std::shared_ptr<Expression> Parser::equality()
 {
-    Expression *result = this->comparison();
+    std::shared_ptr<Expression> result = this->comparison();
     while (this->match_any({{ TOKEN_BANG_EQUAL, TOKEN_EQUAL_EQUAL }})) {
         Token op = PREVIOUS();
-        Expression *expr = this->comparison();
+        std::shared_ptr<Expression> expr = this->comparison();
         result = NEW_NODE(Binary, result, op, expr);
     }
     return result;
@@ -303,12 +303,12 @@ Expression *Parser::equality()
 /*
 logical_and -> equality ("and" equality)*;
 */
-Expression *Parser::logical_and()
+std::shared_ptr<Expression> Parser::logical_and()
 {
-    Expression *result = this->equality();
+    std::shared_ptr<Expression> result = this->equality();
     while (this->match(TOKEN_AND)) {
         Token op = PREVIOUS();
-        Expression *expr = this->equality();
+        std::shared_ptr<Expression> expr = this->equality();
         result = NEW_NODE(Logical, result, op, expr);
     }
     return result;
@@ -317,12 +317,12 @@ Expression *Parser::logical_and()
 /*
 logical_or -> logical_and ("or" logical_and)*;
 */
-Expression *Parser::logical_or()
+std::shared_ptr<Expression> Parser::logical_or()
 {
-    Expression *result = this->logical_and();
+    std::shared_ptr<Expression> result = this->logical_and();
     while (this->match(TOKEN_AND)) {
         Token op = PREVIOUS();
-        Expression *expr = this->logical_and();
+        std::shared_ptr<Expression> expr = this->logical_and();
         result = NEW_NODE(Logical, result, op, expr);
     }
     return result;
@@ -331,11 +331,11 @@ Expression *Parser::logical_or()
 /*
 range -> logical_or ((".." | "...") logical_or)*;
 */
-Expression *Parser::range()
+std::shared_ptr<Expression> Parser::range()
 {
-    Expression *result = this->logical_or();
+    std::shared_ptr<Expression> result = this->logical_or();
     while (this->match_any({{ TOKEN_DOUBLE_DOT, TOKEN_TRIPLE_DOT }})) {
-        Expression *right = this->expression();
+        std::shared_ptr<Expression> right = this->expression();
         result = NEW_NODE(Range, result, right, PREVIOUS().type == TOKEN_DOUBLE_DOT ? false : true);
     }
     return result;
@@ -344,11 +344,11 @@ Expression *Parser::range()
 /*
 assignment -> range ("=" range)*;
 */
-Expression *Parser::assignment()
+std::shared_ptr<Expression> Parser::assignment()
 {
-    Expression *result = this->range();
+    std::shared_ptr<Expression> result = this->range();
     while (this->match(TOKEN_EQUAL)) {
-        Expression *expr = this->range();
+        std::shared_ptr<Expression> expr = this->range();
         result = NEW_NODE(Assign, result, expr);
     }
     return result;
@@ -357,7 +357,7 @@ Expression *Parser::assignment()
 /*
 expression -> assignment;
 */
-Expression *Parser::expression()
+std::shared_ptr<Expression> Parser::expression()
 {
     return this->assignment();
 }
@@ -365,12 +365,12 @@ Expression *Parser::expression()
 /*
 variable_declaration -> IDENTIFIER ":" ("=" expression)?;
 */
-Statement *Parser::variable_declaration()
+std::shared_ptr<Statement> Parser::variable_declaration()
 {
     std::string variable = this->consume(TOKEN_IDENTIFIER, "Expected an identifier in a declaration statement")->to_string();
     this->consume(TOKEN_COLON, "Expected ':' after identifier in a declaration statement");
-    Type *type = this->type();
-    Expression *initializer = nullptr;
+    std::shared_ptr<Type> type = this->type();
+    std::shared_ptr<Expression> initializer;
     if (this->match(TOKEN_EQUAL)) initializer = this->expression();
     return NEW_NODE(Declaration, variable, type, initializer);
 }
@@ -378,29 +378,29 @@ Statement *Parser::variable_declaration()
 /*
 expression_statement -> expression;
 */
-Statement *Parser::expression_statement()
+std::shared_ptr<Statement> Parser::expression_statement()
 {
-    Expression *expr = this->expression();
+    std::shared_ptr<Expression> expr = this->expression();
     return NEW_NODE(ExpressionStatement, expr);
 }
 
 /*
 use_declaration -> "use" IDENTIFIER ("," IDENTIFIER)* "from" STRING;
 */
-Statement *Parser::use_declaration()
+std::shared_ptr<Statement> Parser::use_declaration()
 {
     std::vector<std::string> targets;
     targets.push_back(this->consume(TOKEN_IDENTIFIER, "Expected an identifier after 'use'")->to_string());
     while (this->match(TOKEN_COMMA)) targets.push_back(this->consume(TOKEN_IDENTIFIER, "Expected an identifier after ','")->to_string());
     this->consume(TOKEN_FROM, "Expected 'from' after the import target");
     std::string module = this->consume(TOKEN_STRING, "Expected an identifier after 'from'")->to_string();
-    Parser::format_path(&module, this->file);
-    Use *use;
+    Parser::format_path(module, this->file);
+    std::shared_ptr<Use> use;
     // Parse the contents of the target.
     if (parsed_files.find(module) == parsed_files.end()) {
-        use = NEW_NODE(Use, targets, new std::string(module));
+        use = NEW_NODE(Use, targets, std::make_shared<std::string>(module));
         PREVENT_CYCLIC(use->module);
-        use->code = new std::vector<Statement *>;
+        use->code = std::make_shared<std::vector<std::shared_ptr<Statement>>>();
         Parser(use->module).parse(use->code);
     } else {
         use = NEW_NODE(Use, targets, parsed_files[module].second);
@@ -413,9 +413,9 @@ Statement *Parser::use_declaration()
 /*
 export_declaration -> "export" top_level_declaration
 */
-Statement *Parser::export_declaration()
+std::shared_ptr<Statement> Parser::export_declaration()
 {
-    Statement *stmt = this->top_level_declaration();
+    std::shared_ptr<Statement> stmt = this->top_level_declaration();
     if (stmt->rule == RULE_EXPORT) {
         ADD_LOG("Can't export an export. Does that even make sense?.");
         exit(logger->crash());
@@ -427,22 +427,22 @@ Statement *Parser::export_declaration()
 fun_declaration -> "fun" IDENTIFIER "(" parameters? ")" (":" type)? ("->" expression "\n" | "=>" statement | "{" "\n" statement* "}" "\n");
 parameters -> variable_declaration ("," variable_declaration)*;
 */
-Statement *Parser::fun_declaration()
+std::shared_ptr<Statement> Parser::fun_declaration()
 {
     std::string name = this->consume(TOKEN_IDENTIFIER, "Expected an identifier (function name) after 'fun'.")->to_string();
     this->consume(TOKEN_LEFT_PAREN, "Expected '(' after the function name.");
-    std::vector<Declaration *> parameters;
+    std::vector<std::shared_ptr<Declaration>> parameters;
     if (!this->match(TOKEN_RIGHT_PAREN)) {
         this->parameters(&parameters);
         this->consume(TOKEN_RIGHT_PAREN, "Expected ')' after the function parameters");
     }
-    Type *return_type = nullptr;
+    std::shared_ptr<Type> return_type;
     if (this->match(TOKEN_COLON)) return_type = this->type(false);
-    std::vector<Statement *> body;
+    std::vector<std::shared_ptr<Statement> > body;
     if (this->match(TOKEN_RIGHT_ARROW)) {
-        body.push_back(NEW_NODE(Return, this->expression()));
+        body.push_back(std::move(NEW_NODE(Return, this->expression())));
     } else if (this->match(TOKEN_BIG_RIGHT_ARROW)) {
-        body.push_back(this->statement(false));
+        body.push_back(std::move(this->statement(false)));
     } else if (this->match(TOKEN_LEFT_BRACE)) {
         EXPECT_NEW_LINE();
         body = this->body();
@@ -457,40 +457,40 @@ Statement *Parser::fun_declaration()
 /*
 print_statement -> "print" expression;
 */
-Statement *Parser::print_statement()
+std::shared_ptr<Statement> Parser::print_statement()
 {
-    Expression *expr = this->expression();
+    std::shared_ptr<Expression> expr = this->expression();
     return NEW_NODE(Print, expr);
 }
 
 /*
 if_statement -> "if" expression ("=>" statement | "{" "\n" statement* "}");
 */
-Statement *Parser::if_statement()
+std::shared_ptr<Statement> Parser::if_statement()
 {
-    Expression *condition = this->expression();
-    std::vector<Statement *> then_branch, else_branch;
+    std::shared_ptr<Expression> condition = this->expression();
+    std::vector<std::shared_ptr<Statement> > then_branch, else_branch;
     // Then branch
     if (this->match(TOKEN_LEFT_BRACE)) {
         EXPECT_NEW_LINE();
         then_branch = this->body();
         this->consume(TOKEN_RIGHT_BRACE, "Expected '}' after 'if' body.");
     } else if (this->match(TOKEN_BIG_RIGHT_ARROW)) {
-        then_branch.push_back(this->statement(false));
+        then_branch.push_back(std::move(this->statement(false)));
     } else {
         ADD_LOG("Expected '{' or '=>' after 'if' condition.");
         exit(logger->crash());
     }
     // Else branch
     if (this->match(TOKEN_ELIF)) {
-        else_branch.push_back(this->if_statement());
+        else_branch.push_back(std::move(this->if_statement()));
     } else if (this->match(TOKEN_ELSE)) {
         if (this->match(TOKEN_LEFT_BRACE)) {
             EXPECT_NEW_LINE();
             else_branch = this->body();
             this->consume(TOKEN_RIGHT_BRACE, "Expected '}' after 'else' body.");
         } else if (this->match(TOKEN_BIG_RIGHT_ARROW)) {
-            else_branch.push_back(this->statement(false));
+            else_branch.push_back(std::move(this->statement(false)));
         } else {
             ADD_LOG("Expected '{' or '=>' after 'else'.");
             exit(logger->crash());
@@ -499,16 +499,16 @@ Statement *Parser::if_statement()
     return NEW_NODE(If, condition, then_branch, else_branch);
 }
 
-Statement *Parser::while_statement()
+std::shared_ptr<Statement> Parser::while_statement()
 {
-    Expression *condition = this->expression();
-    std::vector<Statement *> body;
+    std::shared_ptr<Expression> condition = this->expression();
+    std::vector<std::shared_ptr<Statement> > body;
     if (this->match(TOKEN_LEFT_BRACE)) {
         EXPECT_NEW_LINE();
         body = this->body();
         this->consume(TOKEN_RIGHT_BRACE, "Expected '}' after 'while' body.");
     } else if (this->match(TOKEN_BIG_RIGHT_ARROW)) {
-        body.push_back(this->statement(false));
+        body.push_back(std::move(this->statement(false)));
     } else {
         ADD_LOG("Expected '{' or '=>' after 'while' condition.");
         exit(logger->crash());
@@ -519,21 +519,21 @@ Statement *Parser::while_statement()
 /*
 for_statement -> "for" IDENTIFIER ("," IDENTIFIER)? "in" expression ("=>" statement | "{" "\n" statement* "}" "\n");
 */
-Statement *Parser::for_statement()
+std::shared_ptr<Statement> Parser::for_statement()
 {
     std::string index, variable = this->consume(TOKEN_IDENTIFIER, "Expected identifier after 'for'")->to_string();
     if (this->match(TOKEN_COMMA)) {
         index = this->consume(TOKEN_IDENTIFIER, "Expected identifier as the optional second identifier in 'for'")->to_string();
     }
     this->consume(TOKEN_IN, "Expected 'in' after 'for' identifier/s.");
-    Expression *iterator = this->expression();
-    std::vector<Statement *> body;
+    std::shared_ptr<Expression> iterator = this->expression();
+    std::vector<std::shared_ptr<Statement> > body;
     if (this->match(TOKEN_LEFT_BRACE)) {
         EXPECT_NEW_LINE();
         body = this->body();
         this->consume(TOKEN_RIGHT_BRACE, "Expected '}' after 'for' body.");
     } else if (this->match(TOKEN_BIG_RIGHT_ARROW)) {
-        body.push_back(this->statement(false));
+        body.push_back(std::move(this->statement(false)));
     } else {
         ADD_LOG("Expected '{' or '=>' after 'for' iterator.");
         exit(logger->crash());
@@ -544,20 +544,20 @@ Statement *Parser::for_statement()
 /*
 return_statement -> "return" expression?;
 */
-Statement *Parser::return_statement()
+std::shared_ptr<Statement> Parser::return_statement()
 {
     if (this->match_any({{ TOKEN_NEW_LINE, TOKEN_EOF }})) {
         return NEW_NODE(Return, nullptr);
     }
-    Expression *expr = this->expression();
+    std::shared_ptr<Expression> expr = this->expression();
     return NEW_NODE(Return, expr);
 }
 
-Statement *Parser::class_statement()
+std::shared_ptr<Statement> Parser::class_statement()
 {
     std::string name = this->consume(TOKEN_IDENTIFIER, "Expected identifier after 'class'.")->to_string();
     this->consume(TOKEN_LEFT_BRACE, "Expected '{' after 'class' name.");
-    std::vector<Statement *> body = this->class_body();
+    std::vector<std::shared_ptr<Statement> > body = this->class_body();
     this->consume(TOKEN_RIGHT_BRACE, "Expected '}' after 'class' body.");
     return NEW_NODE(Class, name, body);
 }
@@ -571,9 +571,9 @@ statement -> variable_declaration "\n"?
     | print_statement "\n"
     | expression_statement "\n"?;
 */
-Statement *Parser::statement(bool new_line)
+std::shared_ptr<Statement> Parser::statement(bool new_line)
 {
-    Statement *result;
+    std::shared_ptr<Statement> result;
     // Remove blank lines
     while (this->match(TOKEN_NEW_LINE));
 
@@ -620,9 +620,9 @@ top_level_declaration -> use_declaration "\n"
     | class_declaration "\n"
     | fun_declaration "\n";
 */
-Statement *Parser::top_level_declaration()
+std::shared_ptr<Statement> Parser::top_level_declaration()
 {
-    Statement *result;
+    std::shared_ptr<Statement> result;
     // Remove blank lines
     while (this->match(TOKEN_NEW_LINE));
 
@@ -656,9 +656,9 @@ Statement *Parser::top_level_declaration()
     return result;
 }
 
-Statement *Parser::class_body_declaration()
+std::shared_ptr<Statement> Parser::class_body_declaration()
 {
-    Statement *result = nullptr;
+    std::shared_ptr<Statement> result = nullptr;
     // Remove blank lines
     while (this->match(TOKEN_NEW_LINE));
 
@@ -685,44 +685,44 @@ Statement *Parser::class_body_declaration()
     return result;
 }
 
-void Parser::parameters(std::vector<Declaration *> *dest)
+void Parser::parameters(std::vector<std::shared_ptr<Declaration>> *dest)
 {
     if (!CHECK(TOKEN_RIGHT_PAREN)) {
         do {
-            Statement *parameter = this->statement(false);
+            std::shared_ptr<Statement> parameter = this->statement(false);
             if (parameter->rule != RULE_DECLARATION) {
                 ADD_LOG_PAR(parameter->line, parameter->column, "Invalid argument when defining the function. Expected a variable declaration.");
                 exit(logger->crash());
             }
-            dest->push_back(static_cast<Declaration *>(parameter));
+            dest->push_back(std::move(std::static_pointer_cast<Declaration>(parameter)));
         } while (this->match(TOKEN_COMMA));
     }
 }
 
-std::vector<Expression *> Parser::arguments()
+std::vector<std::shared_ptr<Expression> > Parser::arguments()
 {
-    std::vector<Expression *> arguments;
+    std::vector<std::shared_ptr<Expression> > arguments;
     if (!CHECK(TOKEN_RIGHT_PAREN)) {
-        do arguments.push_back(this->expression());
+        do arguments.push_back(std::move(this->expression()));
         while (this->match(TOKEN_COMMA));
     }
     return arguments;
 }
 
-std::vector<Statement *> Parser::body()
+std::vector<std::shared_ptr<Statement> > Parser::body()
 {
-    std::vector<Statement *> body;
+    std::vector<std::shared_ptr<Statement> > body;
     while (!IS_AT_END() && !CHECK(TOKEN_RIGHT_BRACE)) {
-        body.push_back(this->statement());
+        body.push_back(std::move(this->statement()));
     }
     return body;
 }
 
-std::vector<Statement *> Parser::class_body()
+std::vector<std::shared_ptr<Statement> > Parser::class_body()
 {
-    std::vector<Statement *> body;
+    std::vector<std::shared_ptr<Statement> > body;
     while (!IS_AT_END() && !CHECK(TOKEN_RIGHT_BRACE)) {
-        body.push_back(this->class_body_declaration());
+        body.push_back(std::move(this->class_body_declaration()));
     }
     return body;
 }
@@ -733,34 +733,35 @@ type -> "[" type "]"
     | "(" type ("," type)* ("->" type)? ")"
     | IDENTIFIER;
 */
-Type *Parser::type(bool optional)
+std::shared_ptr<Type> Parser::type(bool optional)
 {
-    Type *type;
     if (this->match(TOKEN_LEFT_SQUARE)) {
         // List type.
-        type = new Type(VALUE_LIST, this->type());
+        std::shared_ptr<Type> inner_type = this->type();
+        std::shared_ptr<Type> type = std::make_shared<Type>(VALUE_LIST, inner_type);
         this->consume(TOKEN_RIGHT_SQUARE, "A list type needs to end with ']'");
         return type;
     } else if (this->match(TOKEN_LEFT_BRACE)) {
         // Dict type.
-        type = new Type(VALUE_DICT, this->type());
+        std::shared_ptr<Type> inner_type = this->type();
+        std::shared_ptr<Type> type = std::make_shared<Type>(VALUE_DICT, inner_type);
         this->consume(TOKEN_RIGHT_BRACE, "A dictionary type needs to end with '}'");
         return type;
     } else if (this->match(TOKEN_LEFT_PAREN)) {
         // Function type.
         if (this->match(TOKEN_RIGHT_PAREN)) {
             // Empty function (no paramenters or return type)
-            return new Type(VALUE_FUN);
+            return std::make_shared<Type>(VALUE_FUN);
         } else if (this->match(TOKEN_RIGHT_ARROW)) {
             // Function without arguments and only a return type.
-            Type *t = this->type();
+            std::shared_ptr<Type> type = this->type();
             this->consume(TOKEN_RIGHT_PAREN, "A function type needs to end with ')'");
-            return new Type(VALUE_FUN, t);
+            return std::make_shared<Type>(VALUE_FUN, type);
         }
         // The function have arguments (+ return)?
-        type = new Type(VALUE_FUN);
+        std::shared_ptr<Type> type = std::make_shared<Type>(VALUE_FUN);
         // Add the parameters.
-        do type->parameters.push_back(this->type());
+        do type->parameters.push_back(std::move(this->type()));
         while (this->match(TOKEN_COMMA));
         // Add the return type if needed.
         if (this->match(TOKEN_RIGHT_ARROW)) type->inner_type = this->type();
@@ -768,7 +769,8 @@ Type *Parser::type(bool optional)
         return type;
     } else if (CHECK(TOKEN_IDENTIFIER)) {
         // Other types (native + custom).
-        return new Type(this->consume(TOKEN_IDENTIFIER, "Expected an identifier as a type.")->to_string());
+        std::string type = this->consume(TOKEN_IDENTIFIER, "Expected an identifier as a type.")->to_string();
+        return std::make_shared<Type>(type);
     } else if (optional && (CHECK(TOKEN_NEW_LINE) || CHECK(TOKEN_EQUAL))) return nullptr;
 
     ADD_LOG("Unknown type token expected.");
@@ -778,39 +780,51 @@ Type *Parser::type(bool optional)
 /*
 program -> top_level_declaration*;
 */
-void Parser::parse(std::vector<Statement *> *code)
+void Parser::parse(std::shared_ptr<std::vector<std::shared_ptr<Statement>>> &code)
 {
-    std::vector<Token> *tokens = new std::vector<Token>;
-    Lexer *lexer = new Lexer(this->file);
+    printf("----> Parser\n");
+    // Add the file we are going to parse to the file_stack.
+    file_stack.push_back(this->file);
+    // Prepare the token list.
+    std::unique_ptr<std::vector<Token>> tokens = std::make_unique<std::vector<Token>>();
+    Lexer lexer = Lexer(this->file);
     // Scan the tokens.
-    lexer->scan(tokens);
+    lexer.scan(tokens);
     // Token::debug_tokens(*tokens);
-    this->current = &(*tokens)[0];
-    while (!IS_AT_END()) code->push_back(this->top_level_declaration());
+    this->current = &tokens->at(0);
+    while (!IS_AT_END()) code->push_back(std::move(this->top_level_declaration()));
     // Check the code size to avoid empty files.
     if (code->size() == 0) {
         logger->add_entity(this->file, 0, 0, "Empty file detected, you might need to check the file or make sure it's not empty.");
         exit(logger->crash());
     }
     parsed_files[*this->file] = std::make_pair(code, this->file);
-    delete lexer;
+    printf("----> !Parser\n");
 }
 
 Parser::Parser(const char *file)
 {
+    printf("1\n");
     std::string source = std::string(file);
-    Parser::format_path(&source);
-    this->file = new std::string(source);
+    printf("2 %s\n", source.c_str());
+    Parser::format_path(source);
+    printf("3\n");
+    this->file = std::move(std::make_shared<const std::string>(std::string(source)));
+    printf("4\n");
 }
 
-void Parser::format_path(std::string *path, const std::string *parent)
+void Parser::format_path(std::string &path, const std::shared_ptr<const std::string> &parent)
 {
+    printf("Path: %s, parent: %s\n", path.c_str(), parent ? parent->c_str() : "No parent");
     // Add the final .nu if needed.
-    if (path->substr(path->length() - 3) != ".nu") path->append(".nu");
+    if (path.substr(path.length() - 3) != ".nu") path.append(".nu");
+    printf("A\n");
     // Join the parent if needed.
-    std::filesystem::path f;
-    if (parent) f = std::filesystem::path(*parent).remove_filename().string() + *path;
-    else f = *path;
+    std::filesystem::path f = "C:/";
+    printf("B\n");
+    if (parent) f = std::filesystem::path(*parent).remove_filename().string() + path;
+    else {  printf("Sup!\n"); f = path; }
+    printf("C\n");
     /*
     printf(
         "Path: %s\nRelative: %ls\nAbsolute: %ls\nNoFilename: %ls\n",
@@ -820,7 +834,7 @@ void Parser::format_path(std::string *path, const std::string *parent)
         std::filesystem::absolute(f).remove_filename().c_str()
     );
     */
-   *path = std::filesystem::absolute(f).string();
+    path = std::filesystem::absolute(f).string();
 }
 
 #undef CURRENT
