@@ -14,7 +14,7 @@
 // - Argument type match
 // - Assignment type match
 // - List / Dictionary index type
-// - Variable lifetime (last_use)
+//! - Variable lifetime (last_use)
 // - Use / Export declarations.
 // - Check for iterator (must be list / dict)
 // Optimizes:
@@ -26,8 +26,8 @@ class Analyzer
     public:
         // Creates an analyzer given the file name.
         Analyzer(const char *file);
-        // Analyzes the code. The result AST is stored in the destination. It returns the main module.
-        Block analyze(std::vector<Statement *> *destination);
+        // Analyzes the code. The result AST is stored in the destination. It returns the main block.
+        std::shared_ptr<Block> analyze(std::shared_ptr<std::vector<std::shared_ptr<Statement>>> &destination);
 };
 
 #endif
